// sourcecode/BTrak.Packages/snova-dashboard-module/Dashboard-module-components/dashboard-components/src/lib/snovasys-dashboard-management/models/dashboard.model.ts
export class Dashboard {
    cols: number;
    rows: number;
    y: number;
    x: number;
    component: any;
    persistanceJson: any;
    name: string;
    isCustomWidget: any;
    customWidgetId: string;
    dashboardId: string;
    dashboardName: string;
}
export class dashboardModuleinfo {
    modules: any;
}